move prefix settings to appparams
