+++ app/params/app_settings.go
@@ -1,6 +1,8 @@
 package params
 
 import (
+	"log"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -24,3 +26,15 @@ var (
 	// MinMinGasPrice is the minimum value a validator can set for `minimum-gas-prices` his app.toml config
 	MinMinGasPrice = sdk.NewDecCoinFromDec(BondDenom, sdk.MustNewDecFromStr("0.05"))
 )
+
+func init() {
+	// XXX: If other upstream or external application's depend on any of Umee's
+	// CLI or command functionality, then this would require us to move the
+	// SetAddressConfig call to somewhere external such as the root command
+	// constructor and anywhere else we contract the app.
+	SetAddressConfig()
+
+	if AccountAddressPrefix != Name {
+		log.Fatal("AccountAddresPrefix must equal Name")
+	}
+}
