+++ app/params/prefix.go
@@ -1,4 +1,4 @@
-package app
+package params
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
